部署上线

+++ app/admin/internal/logic/admin_info/admin_info.go
@@ -42,7 +42,7 @@ func Login(ctx context.Context, name, password string) (token string, expire tim
 	}
 
 	// 5. 生成JWT Token
-	return utility.GenerateToken(admin.Id)
+	return utility.GenerateToken(uint32(admin.Id))
 }
 
 // Register 管理员注册
